fixed areInvoiceExtractionsCompleted called multiple times instead every 5 seconds
